testkit: Move generators and mock servers to `testkit` package.
