+++ testkit/src/main/scala/ch/epfl/scala/bsp/testkit/mock/AbstractMockServer.scala
@@ -1,8 +1,8 @@
-package ch.epfl.scala.bsp.mock
+package ch.epfl.scala.bsp.testkit.mock
 
 import ch.epfl.scala.bsp
 import ch.epfl.scala.bsp._
-import ch.epfl.scala.bsp.mock.mockServers.BspResponse
+import mockServers.BspResponse
 import io.circe.Json
 import io.circe.syntax._
 import monix.eval.Task
